requirements.txt
# Core dependencies
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Database connections
influxdb-client>=1.38.0
asyncpg>=0.29.0
aiomysql>=0.2.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Log processing
watchdog>=3.0.0
python-dateutil>=2.8.0

# Web framework for API and Dashboard
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6
PyJWT>=2.8.0

# Configuration and validation
pyyaml>=6.0
python-dotenv>=1.0.0

# Notifications
requests>=2.31.0

# Testing (for property-based testing)
hypothesis>=6.82.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0